+++ TestData/VBToCSCharacterization/ConvertSolution/VbNetStandardLib/Class1.cs
@@ -1,0 +1,6 @@
+namespace VbNetStandardLib
+{
+    public class Class1
+    {
+    }
+}